+++ tests/testthat/test-response_subgroups.R
@@ -235,3 +235,25 @@ testthat::test_that("d_rsp_subgroups_colvars functions as expected with valid in
   res <- testthat::expect_silent(result)
   testthat::expect_snapshot(res)
 })
+
+testthat::test_that("tabulate_rsp_subgroups na_str argument works as expected", {
+  adrs <- adrs_200
+
+  df <- extract_rsp_subgroups(
+    variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "STRATA2")),
+    data = adrs,
+    conf_level = 0.95,
+    method = "chisq"
+  )
+  df$or$or[2:5] <- NA
+
+  result <- basic_table() %>%
+    tabulate_rsp_subgroups(
+      df = df,
+      vars = c("n", "prop", "n_tot", "or", "ci", "pval"),
+      na_str = "<No data>"
+    )
+
+  res <- testthat::expect_silent(result)
+  testthat::expect_snapshot(res)
+})